Revert CLI move and lazy imports, keeping only the style() changes
